Admin cannot delete the only admin from a company

+++ src/app/pages/corporate-profile-edit/corporate-profile-edit.component.ts
@@ -48,6 +48,7 @@ export class CorporateProfileEditComponent implements OnInit {
   creatingNew: boolean = false;
   writeWidth: number = 800;
   isUserAdmin: boolean = false;
+  companyAdminCount: number;
 
   constructor(
     private route: ActivatedRoute,
@@ -172,6 +173,19 @@ export class CorporateProfileEditComponent implements OnInit {
     })
   }
 
+  checkCompanyAdminCount() {
+    let employeeRoleIds = this.currentAccount.userProfileProxies.map((proxy) => proxy.role);
+    for( let roleId of employeeRoleIds){
+      if(roleId) {
+        this.roleService.getRoleByID(roleId).toPromise().then((role) => {
+          if (role.title && role.title == "admin") {
+            this.companyAdminCount = employeeRoleIds.filter((item) => item == roleId).length
+          }
+        })
+      }
+    }
+  }
+
   addEmployee(employeeId) {
     if (!this.isUserAdmin){return;}
 
@@ -186,10 +200,23 @@ export class CorporateProfileEditComponent implements OnInit {
     this.companyService.getCompanyByID(this.route.snapshot.params['id']).toPromise().then((result) => { this.currentAccount.userProfileProxies = result.userProfileProxies; this.refreshEmployees(); }));
   }
 
-  deleteEmployee(proxyId){
+  deleteEmployee(proxyId, proxyRoleId){
     if (!this.isUserAdmin){return;}
-    this.companyUserProxyService.deleteCompanyUserProxy(proxyId).then(() =>
-    this.companyService.getCompanyByID(this.route.snapshot.params['id']).toPromise().then((result) => { this.currentAccount.userProfileProxies = result.userProfileProxies; this.refreshEmployees(); }));
+    if(proxyRoleId){
+      this.roleService.getRoleByID(proxyRoleId).toPromise().then((role) => {
+        if (!role.title || role.title !== "admin" || this.companyAdminCount >= 2) {
+            this.companyUserProxyService.deleteCompanyUserProxy(proxyId).then(() =>
+            this.companyService.getCompanyByID(this.route.snapshot.params['id']).toPromise().then((result) => { this.currentAccount.userProfileProxies = result.userProfileProxies; this.refreshEmployees(); }));
+        }
+        else {
+          console.log("Can't delete only admin")
+        }
+      })
+    } else {
+        this.companyUserProxyService.deleteCompanyUserProxy(proxyId).then(() =>
+        this.companyService.getCompanyByID(this.route.snapshot.params['id']).toPromise().then((result) => { this.currentAccount.userProfileProxies = result.userProfileProxies; this.refreshEmployees(); }));
+    }
+
   }
 
   switchLeadership(employeeId) {
@@ -224,8 +251,9 @@ export class CorporateProfileEditComponent implements OnInit {
     let req = {};
     req[key] = value;
     this.companyUserProxyService.updateCompanyUserProxies(proxyId, req).toPromise().then(() =>
-    // this.companyService.getCompanyByID(this.route.snapshot.params['id']).toPromise().then((result) => { this.currentAccount.userProfileProxies = result.userProfileProxies; this.refreshEmployees(); }))
-    {});
+      this.companyService.getCompanyByID(this.route.snapshot.params['id']).toPromise().then((result) => { this.currentAccount.userProfileProxies = result.userProfileProxies; this.refreshEmployees(); })
+    );
+
   }
 
   refreshEmployees() {
@@ -238,7 +266,8 @@ export class CorporateProfileEditComponent implements OnInit {
         "username": i.userProfile.username,
         "startDate": new Date(i.startDate).toDateString(),
         "endDate": new Date(i.endDate).toDateString(),
-        "stillAffiliated": i.stillAffiliated
+        "stillAffiliated": i.stillAffiliated,
+        "role": i.role
       })
     }
     this.userService.getUsers().then(res => {
@@ -248,6 +277,7 @@ export class CorporateProfileEditComponent implements OnInit {
         }).includes(user._id)
       })
     })
+    this.checkCompanyAdminCount()
   }
 
   checkIfEmployee(): boolean{
